refactor: move token method to utils in evm package
